+++ package.json
@@ -23,7 +23,7 @@
   "engines": {
     "node": ">= 12"
   },
-  "main": "lib/",
+  "main": "server/",
   "types": "types/",
   "scripts": {
     "lint": "semistandard --fix",
@@ -40,13 +40,13 @@
     ]
   },
   "directories": {
-    "lib": "lib"
+    "lib": "server"
   },
   "files": [
     "CHANGELOG.md",
     "LICENSE",
     "README.md",
-    "lib/**",
+    "server/**",
     "client/**",
     "types/**",
     "*.d.ts",
